fixes symfony
