+++ packages/Symfony/SymfonyBundle/EcotoneSymfonyBundle.php
@@ -2,11 +2,14 @@
 
 namespace Ecotone\SymfonyBundle;
 
+use Ecotone\Messaging\Config\ConfigurationException;
 use Ecotone\SymfonyBundle\DependencyInjection\Compiler\AliasExternalReferenceForTesting;
 use Ecotone\SymfonyBundle\DependencyInjection\EcotoneExtension;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
+use Ecotone\SymfonyBundle\DependencyInjection\Compiler\RequiredReferencesCompilerPass;
 
 /**
  * Class IntegrationMessagingBundle
@@ -22,6 +25,7 @@ class EcotoneSymfonyBundle extends Bundle
     {
         parent::build($container);
         $container->addCompilerPass(new AliasExternalReferenceForTesting());
+        $container->addCompilerPass(new RequiredReferencesCompilerPass());
     }
 
     public function getContainerExtension(): ExtensionInterface
